+++ nym-vpn-core/crates/nym-vpn-account-controller/src/state_machine/syncing_state/mod.rs
@@ -116,6 +116,7 @@ impl SyncingState {
         }
         let Some(device) = shared_state.device.clone() else {
             return ErrorState::enter(
+                shared_state,
                 SyncError::Internal("Logged in, but no device keys".into()).into(),
             );
         };
@@ -278,7 +279,7 @@ impl<C: ConnectivityMonitor> AccountControllerStateHandler<C> for SyncingState {
                             Err(e) if e.is_retryable() => {
                                 if self.attempts > MAX_SYNCING_ATTEMPTS {
                                     tracing::debug!("Error trying to get account summary, exhausted retries : {}", e.to_string());
-                                    NextAccountControllerState::NewState(ErrorState::enter(e.into()))
+                                    NextAccountControllerState::NewState(ErrorState::enter(shared_state, e.into()))
                                 } else {
                                     tracing::debug!("Error trying to get account summary, retrying : {}", e.to_string());
                                     NextAccountControllerState::NewState(SyncingState::enter(shared_state, self.attempts + 1))
@@ -286,14 +287,14 @@ impl<C: ConnectivityMonitor> AccountControllerStateHandler<C> for SyncingState {
                             },
                             Err(e) => {
                                 tracing::debug!("Error trying to get account summary, not retrying : {}", e.to_string());
-                                NextAccountControllerState::NewState(ErrorState::enter(e.into()))
+                                NextAccountControllerState::NewState(ErrorState::enter(shared_state, e.into()))
                             },
                         }
                     }
                     Err(e) => {
                         tracing::error!("Failed to join on the syncing task : {e}");
                         if self.attempts > MAX_SYNCING_ATTEMPTS {
-                            NextAccountControllerState::NewState(ErrorState::enter(SyncError::Internal("Failed to join on the syncing task".into()).into()))
+                            NextAccountControllerState::NewState(ErrorState::enter(shared_state, SyncError::Internal("Failed to join on the syncing task".into()).into()))
                         } else {
                             NextAccountControllerState::NewState(SyncingState::enter(shared_state, self.attempts + 1))
                         }
@@ -353,6 +354,9 @@ impl<C: ConnectivityMonitor> AccountControllerStateHandler<C> for SyncingState {
                         shared_state.firewall_active = true;
                         self.syncing_state_handle.abort();
                         return_sender.send(Ok(()));
+                    },
+                    AccountCommand::SetRefreshMode(mode) => {
+                        shared_state.refresh_mode = mode;
                     },
 
                     AccountCommand::Common(common_command) => {