+++ nym-vpn-core/crates/nym-vpn-account-controller/src/state_machine/upgrade_mode_state.rs
@@ -48,19 +48,22 @@ impl UpgradeModeState {
             .await
         {
             Err(err) => {
-                return ErrorState::enter(AccountControllerErrorStateReason::storage_err(
-                    UM_STATE_CONTEXT,
-                    err,
-                ));
+                return ErrorState::enter(
+                    shared_state,
+                    AccountControllerErrorStateReason::storage_err(UM_STATE_CONTEXT, err),
+                );
             }
             // we should only ever enter UpgradeModeState after we have just retrieved and saved
             // an upgrade mode credential, so this branch **theoretically** should be unreachable
             Ok(None) => {
-                return ErrorState::enter(AccountControllerErrorStateReason::Internal {
-                    context: "UpgradeModeState".to_string(),
-                    details: "entered 'UpgradeModeState' with no upgrade mode credentials"
-                        .to_string(),
-                });
+                return ErrorState::enter(
+                    shared_state,
+                    AccountControllerErrorStateReason::Internal {
+                        context: "UpgradeModeState".to_string(),
+                        details: "entered 'UpgradeModeState' with no upgrade mode credentials"
+                            .to_string(),
+                    },
+                );
             }
             Ok(Some((_, Some(expiration)))) => {
                 let until_expiration = expiration - OffsetDateTime::now_utc();
@@ -108,6 +111,7 @@ impl UpgradeModeState {
             .await
         {
             return Err(ErrorState::enter(
+                shared_state,
                 AccountControllerErrorStateReason::storage_err(UM_STATE_CONTEXT, err),
             ));
         }
@@ -185,6 +189,9 @@ impl UpgradeModeState {
             AccountCommand::VpnApiFirewallUp(return_sender) => {
                 shared_state.firewall_active = true;
                 return_sender.send(Ok(()));
+            }
+            AccountCommand::SetRefreshMode(mode) => {
+                shared_state.refresh_mode = mode;
             }
             AccountCommand::Common(common_command) => {
                 common_handler::handle_common_command(common_command, shared_state).await